dmioem: Fix HPE type 199 on theoretical big-endian systems
dmi_print_cpuid() expects a pointer to raw data to interpret as a
CPUID, not to a native integer. If we call DWORD() on the data before
calling dmi_print_cpuid(), and once again in dmi_print_cpuid(), we
end up swapping the bytes twice if running on a big-endian system,
which causes completely wrong CPU ID data to be displayed.

The bug does not happen on x86 because DWORD() is essentially
idempotent on this architecture.

Signed-off-by: Jean Delvare <jdelvare@suse.de>